keras-app/requirements.txt
tensorflow
keras
sklearn
flask
gevent
pillow
requests
opencv-python
